issue #2420 - enable the ConcurrentUpdateTest for all db types
Previously this was flagged as a db2-only test, which helped contribute
to us missing this issue.

Additionally, this test was written to highlight a potential issue but
didn't yet take into account our solution which is to return a 409 for
one of the updates (in the case of a conflict).

The updated test will retry on a tight loop, ensuring that we can
continue to make progress in the face of contentious PUT requests to the
same resource.
